% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{arg_factory}
\alias{arg_factory}
\title{Factory for creating argument types}
\usage{
arg_factory(f.conversion, type, short)
}
\arguments{
\item{f.conversion}{function used to do conversion}

\item{type}{type used in argparse}

\item{short}{short name used in help}
}
\description{
Factory for creating argument types
}

