% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_args.R
\name{arg_params}
\alias{arg_params}
\title{Convert the results of an arg_factory into params for argparse}
\usage{
arg_params(name, arg)
}
\arguments{
\item{arg}{results of arg_factory call}
}
\description{
Convert the results of an arg_factory into params for argparse
}

