% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{arg_factory}
\alias{arg_factory}
\title{Factory for creating argument types}
\usage{
arg_factory(f.conversion, type, short, f.invert = identity, plural = FALSE)
}
\arguments{
\item{f.conversion}{function used to do conversion}

\item{type}{type used in argparse}

\item{short}{short name used in help}

\item{f.invert}{function to invert type for argparse of odd types}

\item{plural}{should this arg take multiple values}
}
\description{
Factory for creating argument types
}

