% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-package.R, R/args.R
\docType{package}
\name{args}
\alias{arg_chr}
\alias{arg_date}
\alias{arg_dbl}
\alias{arg_int}
\alias{arg_lgl}
\alias{args}
\alias{args-package}
\alias{narg_chr}
\alias{narg_date}
\alias{narg_dbl}
\alias{narg_int}
\alias{narg_lgl}
\title{args.}
\usage{
arg_int(help, default = NULL, side.effect = NULL)

narg_int(help, default = NULL, side.effect = NULL)

arg_dbl(help, default = NULL, side.effect = NULL)

narg_dbl(help, default = NULL, side.effect = NULL)

arg_lgl(help, default = NULL, side.effect = NULL)

narg_lgl(help, default = NULL, side.effect = NULL)

arg_chr(help, default = NULL, side.effect = NULL)

narg_chr(help, default = NULL, side.effect = NULL)

arg_date(help, default = NULL, side.effect = NULL)

narg_date(help, default = NULL, side.effect = NULL)
}
\arguments{
\item{help}{help message}

\item{default}{default value}

\item{side.effect}{optional side effect function that the argument value will
be passed into}
}
\description{
args.

The \code{argn_} family allows for multiple arguments to be passed when
delimited by spaces.
}
\seealso{
\code{\link{init_args}} for initializing arguments in a script
}

