% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-package.R, R/args.R
\docType{package}
\name{args}
\alias{arg_chr}
\alias{arg_dbl}
\alias{arg_int}
\alias{arg_lgl}
\alias{args}
\alias{args-package}
\title{args.}
\usage{
arg_int(help, default)

arg_dbl(help, default)

arg_lgl(help, default)

arg_chr(help, default)
}
\arguments{
\item{help}{help message}

\item{default}{default value}
}
\description{
args.

Define an argument of a specific type
}
\seealso{
\code{\link{init_args}} for initializing arguments in a script
}

