% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-package.R, R/args.R
\docType{package}
\name{args}
\alias{arg_chr}
\alias{arg_date}
\alias{arg_dbl}
\alias{arg_int}
\alias{arg_lgl}
\alias{argn_chr}
\alias{argn_dbl}
\alias{argn_int}
\alias{argn_lgl}
\alias{args}
\alias{args-package}
\title{args.}
\usage{
arg_int(help, default)

argn_int(help, default)

arg_dbl(help, default)

argn_dbl(help, default)

arg_lgl(help, default)

argn_lgl(help, default)

arg_chr(help, default)

argn_chr(help, default)

arg_date(help, default)
}
\arguments{
\item{help}{help message}

\item{default}{default value}
}
\description{
args.

The \code{argn_} family allows for multiple arguments to be passed when
delimited by spaces.
}
\seealso{
\code{\link{init_args}} for initializing arguments in a script
}

